requirements.txt
Django==1.8.5
django-filter==0.11.0
djangorestframework==3.2.4
Markdown==2.6.2
wheel==0.24.0
django-rest-auth==0.5.0
django-allauth==0.20.0